Set 'GOGS_WORK_DIR' to CWD in wrapper
Some state is ignoring the 'GOGS_WORK_DIR' anyway putting everything in
the CWD, and otherwing try to put in the (immutable) program directory.

The only deterministic and reliable way to manage the state is to set
`GOGS_WORK_DIR` to CWD, and be sure to launch the program from the state
directory.